% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmers.R
\name{build_kmer_database}
\alias{build_kmer_database}
\title{Build kmer database for classifying 16S rRNA and other gene sequences to
a genus when a kmer size is provided.}
\usage{
build_kmer_database(sequences, genera, kmer_size = 8)
}
\arguments{
\item{sequences}{A vector of reference sequences for which we have
genus-level taxonomic information in the same order as the
value for genera.}

\item{genera}{A vector of genus-level taxonomic information for reference
sequences in the same order as the value for sequences.
Ideally, taxonomic information will be provided back to the
domain level with each level separated by semicolons and no
spaces.}

\item{kmer_size}{The length of the nucleotide word to base our classification
on (default = 8)}
}
\value{
A TBD object containing the genus level conditional probability of
         seeing each kmer in a given genus as well as the genus names
}
\description{
Build kmer database for classifying 16S rRNA and other gene sequences to
a genus when a kmer size is provided.
}
