% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trainset19_db}
\alias{trainset19_db}
\title{RDP training set v19}
\format{
A \code{{phylotypr}} database with 24,642 sequences that has been trained
using 8-mer kmers for the RDP taxonomy from phylum down to genus:
\describe{
\item{conditional_prob}{The conditional probabilities for each genus and kmer (log transformed doubles)}
\item{genera}{Taxonomic string with each level separated with a \verb{;}}
}
}
\source{
\itemize{
\item \href{https://mothur.org/wiki/rdp_reference_files/}{mothur-formatted files}
\item \href{https://mothur.org/blog/2024/RDP-v19-reference-files/}{Description of how mothur-formatted files were generated}
\item \href{https://sourceforge.net/p/rdp-classifier/news/2023/08/rdp-classifier-214-august-2023-released/}{RDP sourceforge page}
}
}
\usage{
trainset19_db
}
\description{
The sequence and taxonomy data for the 24,642 sequences found in the
Ribosomal Database Project's trainset19_072023 training set for use with the
naive Bayesian classifier as implemented in the \code{{phylyotypr}} R package.
Generated using the \code{phylotypr::build_kmer_database()} function:
}
\keyword{datasets}
