% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trainset19_df}
\alias{trainset19_df}
\title{RDP training set v19}
\format{
A data frame with 24,642 rows and 3 columns. Each row represents a different
sequence:
\describe{
\item{id}{Sequence accession identifier}
\item{sequence}{DNA sequence string}
\item{taxonomy}{Taxonomic string with each level separated with a \verb{;}}
}
}
\source{
\itemize{
\item \href{https://mothur.org/wiki/rdp_reference_files/}{mothur-formatted files}
\item \href{https://mothur.org/blog/2024/RDP-v19-reference-files/}{Description of how mothur-formatted files were generated}
\item \href{https://sourceforge.net/p/rdp-classifier/news/2023/08/rdp-classifier-214-august-2023-released/}{RDP sourceforge page}
}
}
\usage{
trainset19_df
}
\description{
The sequence and taxonomy data for the 24,642 sequences found in the
Ribosomal Database Project's trainset19_072023 training set for use with the
naive Bayesian classifier as implemented in the \code{{phylyotypr}} R package.
Originally released by the RDP in July 2023
}
\keyword{datasets}
