% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trainset9_rdp}
\alias{trainset9_rdp}
\alias{trainset9_pds}
\title{RDP training set v9}
\format{
A data frame with 3 columns. Each row represents a different
sequence:
\describe{
\item{id}{Sequence accession identifier}
\item{sequence}{DNA sequence string}
\item{taxonomy}{Taxonomic string with each level separated with a \verb{;}}
}

An object of class \code{data.frame} with 10169 rows and 3 columns.
}
\source{
\itemize{
\item \href{https://mothur.org/wiki/rdp_reference_files/}{mothur-formatted files}
\item \href{https://sourceforge.net/projects/rdp-classifier/files/RDP_Classifier_TrainingData/RDPClassifier_16S_trainsetNo9_rawtrainingdata.zip/download}{RDP sourceforge page} # nolint: line_length_linter
}
}
\usage{
trainset9_rdp

trainset9_pds
}
\description{
The sequence and taxonomy data for the 10,049 sequences found in the
Ribosomal Database Project's trainset9_032012 training set for use with the
naive Bayesian classifier as implemented in the \code{{phylyotypr}} R package.
Originally released by the RDP in September 2012. The \code{rdp} version contains
the same sequences as provided by the official RDP version (9,665 bacterial
and 384 archaeal). The \code{pds} version contains extra eukaryotic sequences
including 119 chloroplasts and mitochondria (10,168 total sequences). See the
mothur reference file page in "Sources" for more information. Be sure to see
the \href{https://github.com/mothur}{mothur GitHub project} for repositories to
obtain R packages containing other taxonomic reference data. For example v19
of this reference is the
\href{https://github.com/mothur/trainset19}{trainset19 package}.
}
\keyword{datasets}
