% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{phylotypr_example}
\alias{phylotypr_example}
\title{Get path to phylotypr example}
\usage{
phylotypr_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\value{
A string indicating path to the file listed in \code{path}. If \code{NULL} is
given then the return value is a vector of file names in the
\verb{extdata/} directory
}
\description{
phylotypr comes bundled with some example files in its \code{inst/extdata}
directory. This function make them easy to access.
}
\examples{
phylotypr_example()
phylotypr_example("miseq_sop.fasta.gz")
}
